+++ App.tsx
@@ -1,63 +1,49 @@
 import React from 'react';
-import {
-  StyleSheet,
-  Image,
-  SafeAreaView
-} from 'react-native';
+import {StyleSheet, Image, SafeAreaView} from 'react-native';
 
-import {NavigationContainer} from '@react-navigation/native'
-import {createBottomTabNavigator} from '@react-navigation/bottom-tabs'
-import Screen1 from './src/screens/Screen1';
+import {NavigationContainer} from '@react-navigation/native';
+import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
+import MainScreen from './src/screens/MainScreen';
 import Screen2 from './src/screens/Screen2';
 import Screen3 from './src/screens/Screen3';
 import Screen4 from './src/screens/Screen4';
 
-import chat from './src/assets/images/bottomTab/chat.png'
-import group from './src/assets/images/bottomTab/group.png'
-import myInfo from './src/assets/images/bottomTab/myInfo.png'
-import friends from './src/assets/images/bottomTab/friends.png'
+import chat from './src/assets/images/bottomTab/chat.png';
+import group from './src/assets/images/bottomTab/group.png';
+import myInfo from './src/assets/images/bottomTab/myInfo.png';
+import friends from './src/assets/images/bottomTab/friends.png';
 
-import {height,width,colors} from './src/constants/Index'
+import {height, width, colors} from './src/constants/Index';
 
-const Tab = createBottomTabNavigator()
-
-
+const Tab = createBottomTabNavigator();
 
 const App = () => {
-  
   return (
     <>
       <NavigationContainer>
         <Tab.Navigator
           tabBarOptions={{
-            labelStyle: {fontSize:18},
+            labelStyle: {fontSize: 18},
             activeTintColor: colors.focusTabBar,
-            inactiveTintColor: 'black', 
-            style: { height: height*70 },
-          }}
-        >
+            inactiveTintColor: 'black',
+            style: {height: height * 70},
+          }}>
           <Tab.Screen
             name="Screen1"
             options={{
               tabBarLabel: '그룹',
-              tabBarIcon: ({ color, size }) => (
-                <Image
-                  style={styles.tinyLogo}
-                  source={group}
-                />
+              tabBarIcon: ({color, size}) => (
+                <Image style={styles.tinyLogo} source={group} />
               ),
             }}
-            component={Screen1}
+            component={MainScreen}
           />
           <Tab.Screen
             name="Screen2"
             options={{
               tabBarLabel: '채팅',
-              tabBarIcon: ({ color, size }) => (
-                <Image
-                  style={styles.tinyLogo}
-                  source={chat}
-                />
+              tabBarIcon: ({color, size}) => (
+                <Image style={styles.tinyLogo} source={chat} />
               ),
             }}
             component={Screen2}
@@ -66,25 +52,22 @@ const App = () => {
             name="Screen3"
             options={{
               tabBarLabel: '친구',
-              tabBarIcon: ({ color, size }) => (
+              tabBarIcon: ({color, size}) => (
                 <Image
-                resizeMode={'cover'}
-                  style={{height:height*25,width:width*40}}
+                  resizeMode={'cover'}
+                  style={{height: height * 25, width: width * 40}}
                   source={friends}
                 />
               ),
             }}
             component={Screen3}
           />
-           <Tab.Screen
+          <Tab.Screen
             name="Screen4"
             options={{
               tabBarLabel: '계정',
-              tabBarIcon: ({ color, size }) => (
-                <Image
-                  style={styles.tinyLogo}
-                  source={myInfo}
-                />
+              tabBarIcon: ({color, size}) => (
+                <Image style={styles.tinyLogo} source={myInfo} />
               ),
             }}
             component={Screen4}
@@ -96,10 +79,9 @@ const App = () => {
 };
 
 const styles = StyleSheet.create({
-  
   tinyLogo: {
-    width: width*20,
-    height: height*20,
-  }
+    width: width * 20,
+    height: height * 20,
+  },
 });
 export default App;
